# add vgg16 performance script