--- a/perf/conv.jl
+++ b/perf/conv.jl
@@ -1,5 +1,5 @@
 for ch in [1, 3, 16, 64]
-    x = rand(64, 64, ch, 64)
+    x = rand(Float32, 64, 64, ch, 64)
     model = Conv((3,3), ch=>ch)
     println("CPU ch=$ch")
     run_benchmark(model, x, cuda=false)