--- a/perf/dense.jl
+++ b/perf/dense.jl
@@ -1,5 +1,5 @@
-for n in [4, 16, 64, 256, 1024]
-    x = rand(n, n)
+for n in [2, 20, 200, 2000]
+    x = randn(Float32, n, n)
     model = Dense(n, n)
     println("CPU n=$n")
     run_benchmark(model, x, cuda=false)