--- a/perf/runbenchmarks.jl
+++ b/perf/runbenchmarks.jl
@@ -1,0 +1,12 @@
+using Flux
+
+include("bench_utils.jl")
+
+@info "Benchmark Dense"
+include("dense.jl")
+
+@info "Benchmark Conv"
+include("conv.jl")
+
+@info "Benchmark VGG"
+include("vgg.jl")