We could start collecting a few (possibly more granular) perf scripts, as a first crude benchmarking facility. 

Related discussion in #1527 .

On my laptop,  performances are the same for Flux master and v0.11.6.

```julia
# Julia 1.6-rc1, Flux v0.11.6, CUDA v2.6.1, Zygote v0.6.3, NNlib v0.7.16
CPU benchmark
  forward
  1.752 s (915 allocations: 288.05 MiB)
  backward
  4.166 s (20840 allocations: 1.01 GiB)
  forw and back
  6.117 s (51836 allocations: 1.63 GiB)
CUDA benchmark
  forward
  21.032 ms (4115 allocations: 156.30 KiB)
  backward
  139.990 ms (9218 allocations: 365.88 KiB)
  forw and back
  168.142 ms (19244 allocations: 2.57 MiB)
```
```julia
# Julia 1.6-rc1, Flux master, CUDA v2.6.1, Zygote v0.6.3, NNlib v0.7.16
CPU benchmark
  forward
  1.682 s (1123 allocations: 220.57 MiB)
  backward
  4.210 s (20153 allocations: 1.08 GiB)
  forw and back
  6.382 s (32100 allocations: 1.69 GiB)
CUDA benchmark
  forward
  21.704 ms (4052 allocations: 149.59 KiB)
  backward
  139.710 ms (9832 allocations: 377.77 KiB)
  forw and back
  169.053 ms (20759 allocations: 2.32 MiB)
```
```julia

```