@darsnack 2021-03-12
This is great! I think the SciML benchmarks reside in a separate repo. I'm not sure if this is a subjective choice or a technical one (@ChrisRackauckas or @DhairyaLGandhi would can comment). Ideally, `ModelZooKeeper` will run these in the future, but even having them in the repo right now will make us all feel better about whether a release is ready.

One change I would make is to structure the directory like:
```plaintext
benchmarks
\_ metalhead_vgg16
   \_ Manifest.toml
   \_ Project.toml
   \_ script.jl
...
\_ run_all.jl
```
Each benchmark gets its own sub-directory with its own environment to be exactly reproducible on everyone's machines (and CI).

If we get this merged in, I can submit another PR with more vision related benchmarks.