@ChrisRackauckas 2021-03-12
It's because the benchmark results can get huge. https://github.com/SciML/SciMLBenchmarks.jl hosts a huge set of PDFs and webpages to display the current benchmarks. You don't want to ship all of that with every install