@DhairyaLGandhi 2021-03-12
I have some work on benchmarks that I have been coordinating already. We will populate a public https://speed.fluxml.ai soon