@darsnack 2021-03-12
Nice, is it possible for us to PR/commit to whatever feeds that website? I'm just thinking to the original motivation for this PR — having one set of benchmarks that we are all referring to as we track whether code is ready to be released.