@CarloLucibello 2021-03-13
> I have some work on benchmarks that I have been coordinating already. We will populate a public https://speed.fluxml.ai soon

That seems cool.

So let's merge this as a short term solution, then we can expand, move somewhere else, ditch entirely when a better solution becomes available.

> One change I would make is to structure the directory like:

I added some other scripts (included a runall) and Project/Manifest, but kept a flat structure. Feel free to modify in the future.